Reorg booking into domains/ (maximal split)
The complex domain — queries, mutations, a race-sensitive acceptance lifecycle,
and the one Google-touching action — split to the full structure:
- domains/booking/tables.ts    — bookingPages / availabilityOverrides / bookings
- domains/booking/model.ts     — constants, page reads, server-side slot grid
- domains/booking/queries.ts   — read handlers (incl. getBookingContext)
- domains/booking/mutations.ts — page/override/request writes + the acceptance
  claim/mark/release lease + expiry continuations
- domains/booking/service.ts   — acceptBooking (the action; Google create + 409
  reconcile + mirror)
- booking.ts (root)            — thin facade, Convex wrappers at the same paths
  with identical arg validators

Behavior-preserving: the Stage-1 booking acceptance race/reconcile itests pass
through the moved handlers, api.d.ts diff is purely additive, and every
api.booking.* / internal.booking.* path (including the queued expireBooking) is
unchanged. Backend 100 unit + 25 integration + types green.
