+++ packages/backend/convex/_generated/api.d.ts
@@ -18,6 +18,11 @@ import type * as calendar from "../calendar.js";
 import type * as calendarSync from "../calendarSync.js";
 import type * as contacts from "../contacts.js";
 import type * as crons from "../crons.js";
+import type * as domains_booking_model from "../domains/booking/model.js";
+import type * as domains_booking_mutations from "../domains/booking/mutations.js";
+import type * as domains_booking_queries from "../domains/booking/queries.js";
+import type * as domains_booking_service from "../domains/booking/service.js";
+import type * as domains_booking_tables from "../domains/booking/tables.js";
 import type * as domains_marketing_mutations from "../domains/marketing/mutations.js";
 import type * as domains_marketing_tables from "../domains/marketing/tables.js";
 import type * as domains_notifications_model from "../domains/notifications/model.js";
@@ -68,6 +73,11 @@ declare const fullApi: ApiFromModules<{
   calendarSync: typeof calendarSync;
   contacts: typeof contacts;
   crons: typeof crons;
+  "domains/booking/model": typeof domains_booking_model;
+  "domains/booking/mutations": typeof domains_booking_mutations;
+  "domains/booking/queries": typeof domains_booking_queries;
+  "domains/booking/service": typeof domains_booking_service;
+  "domains/booking/tables": typeof domains_booking_tables;
   "domains/marketing/mutations": typeof domains_marketing_mutations;
   "domains/marketing/tables": typeof domains_marketing_tables;
   "domains/notifications/model": typeof domains_notifications_model;
